git-svn-id: http://plugins.svn.wordpress.org/2performant-product-importer/branches/1.0@385567 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ toolbox.php
@@ -115,7 +115,7 @@ function TP_AJAX_wrapper_getProducts() {
 		$perBatch = get_option( 'tp_options_add_feed', array('update_batch_size' => 50) );
 		$perBatch = isset($perBatch['update_batch_size']) ? $perBatch['update_batch_size'] : 50;
 		
-		$prods = get_posts( 'type=' . tp_get_post_type() . '&post_status=publish,draft,pending&numberposts='.intval($perBatch).'&offset='.intval($page*$perBatch) );
+		$prods = get_posts( 'type=' . tp_get_post_type() . '&post_status=publish,draft,pending&meta_key=tp_product_ID&numberposts='.intval($perBatch).'&offset='.intval($page*$perBatch) );
 		$ids = array();
 		foreach ( $prods as $k => $v ) {
 			$t = get_post_meta( $v->ID, 'tp_product_info', true );
@@ -190,7 +190,25 @@ function TP_AJAX_wrapper_updateProduct() {
 		if ( ! $p )
 			throw new Exception( sprintf( __( 'Invalid post ID: %1$s' ), $id ) );
 		
+		$product_id = intval( get_post_meta( $id, 'tp_product_ID', true ) );
+		if( !$product_id )
+			throw new Exception( sprintf( __('No product ID attached for post %d'), $id ) );
+			
 		$product = tp_get_post_product_data( $id );
+		if( !is_object($product) || empty($product->{'product-store-id'}) || empty($product->id) ) {
+			$product = tp_get_wrapper()->product_store_products_search('approved',"@id $product_id");
+			if( empty( $product ) )
+				throw new Exception( sprintf( __('Invalid product data for ID %d'), $product_id ) );
+			if( !is_object($product) ) {
+				if( !is_array( $product ) )
+					throw new Exception( sprintf( __('Invalid product data received from API for ID %d'), $product_id ) );
+				while( is_array( $product ) ) {
+					$product = array_pop( $product );
+				}
+				if( !is_object($product) || empty($product->{'product-store-id'}) || empty($product->id) )
+					throw new Exception( sprintf( __('Invalid product data received from API for ID %d'), $product_id ) );
+			}
+		}
 		
 		try {
 			require_once 'api.php';
@@ -203,8 +221,10 @@ function TP_AJAX_wrapper_updateProduct() {
 		}
 		
 		if ( ! $live_product ) {
-			// Delete the post
-			wp_delete_post( $id );
+			if( tp_get_option( 'add_feed', 'trash_expired', true ) ) {
+				// Delete the post
+				wp_delete_post( $id );
+			}
 			throw new Exception( sprintf( __( 'Expired product: %1$s' ), ( $product ? tp_strtopinfo( '%brand% %title% (%id%)', $product ) : '' ) ) );
 		}
